+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * nostr: update `TryIntoUrl::try_into_url` fingerprint ([Yuki Kishimoto])
 * nostr: bump `bitcoin` to `0.31` ([Yuki Kishimoto])
 * sdk: bump `lnurl-pay` to `0.4` ([Yuki Kishimoto])
+* sdk: improve `proxy` options ([Yuki Kishimoto])
 * pool: bump `async-wsocket` to `0.4` ([Yuki Kishimoto])
 * pool: return error if `urls` arg is empty in `InternalRelayPool::get_events_from` ([Yuki Kishimoto])
 * pool: allow to disable `RelayLimits` ([Yuki Kishimoto])